accounts/serializers.py
# we have this bc we want to see all users so we must make a view --> creates serializers
from rest_framework import serializers
from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username',)