requirements.txt
aiohttp==3.7.0
asgiref==3.5.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
bleach==3.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
distlib==0.3.4
Django==2.2.27
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
docutils==0.16
filelock==3.6.0
future==0.18.2
geocoder==1.38.1
geoip2==4.1.0
idna==2.10
ip2geotools==0.1.5
IP2Location==8.5.1
isort==5.6.4
keyring==21.4.0
lazy-object-proxy==1.5.1
lxml==4.6.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.0.0
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pip-review==1.1.0
pipenv==2022.1.8
pkginfo==1.6.0
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycodestyle==2.6.0
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
ratelim==0.1.6
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.1
tqdm==4.50.2
twine==3.2.0
typed-ast==1.4.1
typing==3.7.4.3
urllib3==1.25.11
virtualenv==20.13.3
virtualenv-clone==0.5.7
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.2
